+++ build.gradle
@@ -15,7 +15,7 @@ apply plugin: 'eclipse'
 apply plugin: 'org.springframework.boot'
 apply plugin: 'io.spring.dependency-management'
 
-group = 'com.example'
+group = 'com.jimmy'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = 1.8
