--- src/main/java/com/example/demo/domain/LocalDateTimeConverter.java
@@ -1,19 +1,0 @@
-package com.example.demo.domain;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-
-@Converter(autoApply = true)
-public class LocalDateTimeConverter implements AttributeConverter<LocalDateTime, Timestamp> {
-    @Override
-    public Timestamp convertToDatabaseColumn(LocalDateTime localDateTime) {
-        return localDateTime != null ? Timestamp.valueOf(localDateTime) : null;
-    }
-
-    @Override
-    public LocalDateTime convertToEntityAttribute(Timestamp timestamp) {
-        return timestamp != null ? timestamp.toLocalDateTime() : null;
-    }
-}
