--- src/main/java/com/example/demo/support/domain/UrlGeneratable.java
@@ -1,5 +1,0 @@
-package com.example.demo.support.domain;
-
-public interface UrlGeneratable {
-    String generateUrl();
-}
