--- src/main/java/com/example/demo/web/HomeController.java
@@ -1,16 +1,0 @@
-package com.example.demo.web;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-    public static final Logger logger = LoggerFactory.getLogger(HomeController.class);
-
-    @GetMapping("/")
-    public String index() {
-        return "/index";
-    }
-}
