+++ src/main/java/com/jimmy/crawler/domain/StockRepository.java
@@ -1,0 +1,13 @@
+package com.jimmy.crawler.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface StockRepository extends JpaRepository<Stock, Long> {
+    Stock findByName(String name);
+    List<Stock> findAllByOrderByPriceDesc();
+    List<Stock> findAllByOrderByChangePercentDesc();
+    List<Stock> findAllByOrderByChangePercentAsc();
+    List<Stock> findAllByOrderByPriceAsc();
+}
