+++ src/main/java/com/jimmy/crawler/support/domain/UrlGeneratable.java
@@ -1,0 +1,5 @@
+package com.jimmy.crawler.support.domain;
+
+public interface UrlGeneratable {
+    String generateUrl();
+}
